backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.13.5
blis==1.3.0
catalogue==2.0.10
certifi==2025.8.3
chardet==3.0.4
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.22.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
deep-translator==1.11.4
distro==1.9.0
fastapi==0.117.1
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httpx==0.28.1
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
nltk==3.9.1
numpy==2.3.3
openai==1.109.1
packaging==25.0
preshed==3.0.10
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyPDF2==3.0.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
rfc3986==1.5.0
rich==14.1.0
scikit-learn==1.7.2
scipy==1.16.2
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
soupsieve==2.8
srsly==2.5.1
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.8.0
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
weasel==0.4.1
wrapt==1.17.3
